+++ src/C_PETSc_Interfaces.F90
@@ -214,9 +214,10 @@ module c_petsc_interfaces
 
    interface   
       
-      subroutine create_cf_is_kokkos(A_array, index_fine, index_coarse) &
+      subroutine create_cf_is_kokkos(handle, A_array, index_fine, index_coarse) &
          bind(c, name="create_cf_is_kokkos")
          use iso_c_binding
+         type(c_ptr), value :: handle
          integer(c_long_long) :: A_array
          integer(c_long_long) :: index_fine
          integer(c_long_long) :: index_coarse
@@ -226,9 +227,10 @@ module c_petsc_interfaces
 
    interface   
       
-      subroutine pmisr_kokkos(A_array, max_luby_steps, pmis_int, measure_local, zero_meaure_c_point_int) &
+      subroutine pmisr_kokkos(handle, A_array, max_luby_steps, pmis_int, measure_local, zero_meaure_c_point_int) &
          bind(c, name="pmisr_kokkos")
          use iso_c_binding
+         type(c_ptr) :: handle
          integer(c_long_long) :: A_array
          type(c_ptr), value :: measure_local
          integer(c_int), value :: max_luby_steps, pmis_int, zero_meaure_c_point_int
@@ -238,9 +240,10 @@ module c_petsc_interfaces
 
    interface   
       
-      subroutine MatDiagDomRatio_kokkos(A_array, max_dd_ratio_achieved, local_rows_aff) &
+      subroutine MatDiagDomRatio_kokkos(handle, A_array, max_dd_ratio_achieved, local_rows_aff) &
          bind(c, name="MatDiagDomRatio_kokkos")
          use iso_c_binding
+         type(c_ptr), value :: handle
          integer(c_long_long) :: A_array
          PetscReal :: max_dd_ratio_achieved
          PetscInt :: local_rows_aff
@@ -250,10 +253,11 @@ module c_petsc_interfaces
 
    interface   
       
-      subroutine ddc_kokkos(A_array, fraction_swap, max_dd_ratio, max_dd_ratio_achieved, Aff_array, &
+      subroutine ddc_kokkos(handle, A_array, fraction_swap, max_dd_ratio, max_dd_ratio_achieved, Aff_array, &
             random_numbers_ptr) &
          bind(c, name="ddc_kokkos")
          use iso_c_binding
+         type(c_ptr), value :: handle
          integer(c_long_long) :: A_array
          PetscReal, value :: fraction_swap
          PetscReal, value :: max_dd_ratio
@@ -266,9 +270,10 @@ module c_petsc_interfaces
    
    interface   
       
-      subroutine copy_cf_markers_d2h(cf_markers_local) &
+      subroutine copy_cf_markers_d2h(handle, cf_markers_local) &
          bind(c, name="copy_cf_markers_d2h")
          use iso_c_binding
+         type(c_ptr), value :: handle
          type(c_ptr), value :: cf_markers_local
       end subroutine copy_cf_markers_d2h         
  
@@ -276,9 +281,10 @@ module c_petsc_interfaces
    
    interface   
       
-      subroutine copy_diag_dom_ratio_d2h(diag_dom_ratio_local) &
+      subroutine copy_diag_dom_ratio_d2h(handle, diag_dom_ratio_local) &
          bind(c, name="copy_diag_dom_ratio_d2h")
          use iso_c_binding
+         type(c_ptr), value :: handle
          type(c_ptr), value :: diag_dom_ratio_local
       end subroutine copy_diag_dom_ratio_d2h
  
@@ -286,19 +292,13 @@ module c_petsc_interfaces
 
    interface   
       
-      subroutine delete_device_cf_markers() &
-         bind(c, name="delete_device_cf_markers")
+      ! Destroys the device cf markers and diag dom ratio behind handle
+      ! (created by pmisr_kokkos) and sets it to c_null_ptr
+      subroutine destroy_cf_markers_kokkos(handle) &
+         bind(c, name="destroy_cf_markers_kokkos")
          use iso_c_binding
-      end subroutine delete_device_cf_markers         
- 
-   end interface       
-
-   interface   
-      
-      subroutine delete_device_diag_dom_ratio() &
-         bind(c, name="delete_device_diag_dom_ratio")
-         use iso_c_binding
-      end subroutine delete_device_diag_dom_ratio
+         type(c_ptr) :: handle
+      end subroutine destroy_cf_markers_kokkos
  
    end interface